feature: refactor gscan to use it as app, as lib or as cli
- use ghost-ignition
- create app folder for web usage
