+++ app/public/js/gscan.js
