+++ app/tpl/partials/rule-fail.hbs
