+++ app/tpl/partials/rule-pass.hbs
