+++ app/tpl/partials/toggle.hbs
