+++ lib/checks/005-template-compile.js
@@ -1,6 +1,6 @@
 var _       = require('lodash'),
     hbs     = require('express-hbs'),
-    spec    = require('../../spec.js'),
+    spec    = require('../spec.js'),
     checkTemplatesCompile;
 
 checkTemplatesCompile = function checkTemplatesCompile(theme) {
