+++ lib/format.js
@@ -1,5 +1,5 @@
 var _          = require('lodash'),
-    spec       = require('../spec'),
+    spec       = require('./spec'),
     format,
     calcScore,
     levelWeights = {
