+++ lib/index.js
@@ -3,7 +3,6 @@ var _       = require('lodash'),
     checker = require('./checker'),
     format   = require('./format'),
     readZip = require('./read-zip'),
-
     middleware,
     checkZip;
 
@@ -36,23 +35,8 @@ checkZip = function checkZip(themePath, options) {
     });
 };
 
-middleware = function middleware(req, res, next) {
-    var zip = {
-        path: req.file.path,
-        name: req.file.originalname
-    };
-
-    checkZip(zip).then(function processResult(theme) {
-        pfs.removeDir(req.file.path).then(function () {
-            res.theme = theme;
-            next();
-        });
-    });
-};
-
 module.exports = {
     check: checker,
     checkZip: checkZip,
-    middleware: middleware,
     format: format
 };
