+++ lib/read-zip.js
@@ -1,11 +1,10 @@
-var debug      = require('debug')('gscan:zip'),
+var debug      = require('ghost-ignition').debug('zip'),
     path       = require('path'),
     Promise    = require('bluebird'),
     extract    = require('extract-zip'),
     uuid       = require('node-uuid'),
     _          = require('lodash'),
     pfs        = require('./promised-fs'),
-
     readZip,
     resolveBaseDir;
