+++ lib/spec.js
