+++ package.json
@@ -24,7 +24,7 @@
   "main": "lib",
   "license": "MIT",
   "scripts": {
-    "start": "node app.js",
+    "start": "node app/index.js",
     "test": "mocha"
   },
   "bin": {
@@ -34,14 +34,13 @@
     "bluebird": "3.0.6",
     "chalk": "1.1.1",
     "commander": "2.9.0",
-    "debug": "2.2.0",
     "express": "4.13.3",
     "express-hbs": "0.8.4",
     "extract-zip": "1.3.0",
     "fs-extra": "0.26.2",
+    "ghost-ignition": "0.0.3",
     "lodash": "3.10.1",
     "multer": "1.1.0",
-    "nconf": "0.8.4",
     "node-uuid": "1.4.2",
     "package-json-validator": "0.6.0",
     "require-dir": "0.1.0"
@@ -56,5 +55,10 @@
     "mocha": "2.4.5",
     "shipit-deploy": "2.1.3",
     "should": "7.1.0"
-  }
+  },
+  "files": [
+    "lib",
+    "bin",
+    "app"
+  ]
 }
